MLService/server/requirements.txt
# --extra-index-url https://google-coral.github.io/py-repo/ tflite_runtime
numpy~=1.19.2
pika
python-dotenv
# gcloud
google-cloud-storage
pandas